+++ test/pleroma/web/feed/user_controller_test.exs
@@ -8,7 +8,6 @@ defmodule Pleroma.Web.Feed.UserControllerTest do
   import Pleroma.Factory
   import SweetXml
 
-  alias Pleroma.Config
   alias Pleroma.Object
   alias Pleroma.User
   alias Pleroma.Web.CommonAPI
@@ -20,7 +19,7 @@ defmodule Pleroma.Web.Feed.UserControllerTest do
     setup do: clear_config([:feed])
 
     setup do
-      Config.put(
+      clear_config(
         [:feed, :post_title],
         %{max_length: 15, omission: "..."}
       )
@@ -236,7 +235,7 @@ defmodule Pleroma.Web.Feed.UserControllerTest do
     setup do: clear_config([:instance, :public])
 
     test "returns 404 for user feed", %{conn: conn} do
-      Config.put([:instance, :public], false)
+      clear_config([:instance, :public], false)
       user = insert(:user)
 
       {:ok, _} = CommonAPI.post(user, %{status: "test"})