+++ test/pleroma/web/plugs/rate_limiter_test.exs
@@ -6,7 +6,6 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
   use Pleroma.Web.ConnCase
 
   alias Phoenix.ConnTest
-  alias Pleroma.Config
   alias Pleroma.Web.Plugs.RateLimiter
   alias Plug.Conn
 
@@ -22,8 +21,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
     setup do: clear_config([Pleroma.Web.Plugs.RemoteIp, :enabled])
 
     test "config is required for plug to work" do
-      Config.put([:rate_limit, @limiter_name], {1, 1})
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, @limiter_name], {1, 1})
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       assert %{limits: {1, 1}, name: :test_init, opts: [name: :test_init]} ==
                [name: @limiter_name]
@@ -54,8 +53,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
     scale = 80
     limit = 5
 
-    Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
-    Config.put([:rate_limit, limiter_name], {scale, limit})
+    clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+    clear_config([:rate_limit, limiter_name], {scale, limit})
 
     plug_opts = RateLimiter.init(name: limiter_name)
     conn = build_conn(:get, "/")
@@ -86,8 +85,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
     test "`bucket_name` option overrides default bucket name" do
       limiter_name = :test_bucket_name
 
-      Config.put([:rate_limit, limiter_name], {1000, 5})
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], {1000, 5})
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       base_bucket_name = "#{limiter_name}:group1"
       plug_opts = RateLimiter.init(name: limiter_name, bucket_name: base_bucket_name)
@@ -101,8 +100,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
 
     test "`params` option allows different queries to be tracked independently" do
       limiter_name = :test_params
-      Config.put([:rate_limit, limiter_name], {1000, 5})
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], {1000, 5})
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       plug_opts = RateLimiter.init(name: limiter_name, params: ["id"])
 
@@ -117,8 +116,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
 
     test "it supports combination of options modifying bucket name" do
       limiter_name = :test_options_combo
-      Config.put([:rate_limit, limiter_name], {1000, 5})
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], {1000, 5})
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       base_bucket_name = "#{limiter_name}:group1"
 
@@ -140,8 +139,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
   describe "unauthenticated users" do
     test "are restricted based on remote IP" do
       limiter_name = :test_unauthenticated
-      Config.put([:rate_limit, limiter_name], [{1000, 5}, {1, 10}])
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], [{1000, 5}, {1, 10}])
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       plug_opts = RateLimiter.init(name: limiter_name)
 
@@ -180,8 +179,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
 
       scale = 50
       limit = 5
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
-      Config.put([:rate_limit, limiter_name], [{1000, 1}, {scale, limit}])
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], [{1000, 1}, {scale, limit}])
 
       plug_opts = RateLimiter.init(name: limiter_name)
 
@@ -202,8 +201,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
 
     test "different users are counted independently" do
       limiter_name = :test_authenticated2
-      Config.put([:rate_limit, limiter_name], [{1, 10}, {1000, 5}])
-      Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+      clear_config([:rate_limit, limiter_name], [{1, 10}, {1000, 5}])
+      clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
       plug_opts = RateLimiter.init(name: limiter_name)
 
@@ -232,8 +231,8 @@ defmodule Pleroma.Web.Plugs.RateLimiterTest do
 
   test "doesn't crash due to a race condition when multiple requests are made at the same time and the bucket is not yet initialized" do
     limiter_name = :test_race_condition
-    Pleroma.Config.put([:rate_limit, limiter_name], {1000, 5})
-    Pleroma.Config.put([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
+    clear_config([:rate_limit, limiter_name], {1000, 5})
+    clear_config([Pleroma.Web.Endpoint, :http, :ip], {8, 8, 8, 8})
 
     opts = RateLimiter.init(name: limiter_name)